+++ src/main/java/com/shy_polarbear/server/domain/image/dto/request/DeleteImageRequest.java
@@ -1,11 +1,9 @@
-package com.shy_polarbear.server.domain.images.dto.request;
+package com.shy_polarbear.server.domain.image.dto.request;
 
-
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
-import org.springframework.web.multipart.MultipartFile;
+import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.List;
@@ -13,13 +11,11 @@ import java.util.List;
 import static com.shy_polarbear.server.global.common.constants.BusinessLogicConstants.MAX_FEED_IMAGES_COUNT;
 import static com.shy_polarbear.server.global.common.constants.BusinessLogicConstants.MIN_IMAGE_COUNT;
 
+@NoArgsConstructor
+@AllArgsConstructor
 @Getter
-@Setter
-public class UploadImageRequest {
-
-    @NotBlank
-    private String type;
+public class DeleteImageRequest {
     @NotNull
     @Size(min = MIN_IMAGE_COUNT, max = MAX_FEED_IMAGES_COUNT)
-    private List<MultipartFile> imageFiles;
+    private List<String> imageUrls;
 }
