+++ src/main/java/com/shy_polarbear/server/domain/image/dto/request/UploadImageRequest.java
@@ -1,4 +1,5 @@
-package com.shy_polarbear.server.domain.images.dto.request;
+package com.shy_polarbear.server.domain.image.dto.request;
+
 
 import lombok.Getter;
 import lombok.Setter;
@@ -12,15 +13,13 @@ import java.util.List;
 import static com.shy_polarbear.server.global.common.constants.BusinessLogicConstants.MAX_FEED_IMAGES_COUNT;
 import static com.shy_polarbear.server.global.common.constants.BusinessLogicConstants.MIN_IMAGE_COUNT;
 
-@Setter
 @Getter
-public class UpdateImageRequest {
+@Setter
+public class UploadImageRequest {
+
     @NotBlank
     private String type;
     @NotNull
     @Size(min = MIN_IMAGE_COUNT, max = MAX_FEED_IMAGES_COUNT)
-    private List<MultipartFile> newImageFiles;
-    @NotNull
-    @Size(min = MIN_IMAGE_COUNT, max = MAX_FEED_IMAGES_COUNT)
-    private List<String> oldImageUrls;
+    private List<MultipartFile> imageFiles;
 }
