+++ src/main/java/com/shy_polarbear/server/domain/image/exception/ImageException.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.images.exception;
+package com.shy_polarbear.server.domain.image.exception;
 
 import com.shy_polarbear.server.global.exception.CustomException;
 import com.shy_polarbear.server.global.exception.ExceptionStatus;
